requirements.txt
streamlit
numpy
torch
transformers
torchvision
pillow
gdown